+++ packages/vue/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { default as CmButton } from './components/button/CmButton.vue';
+export type { CmButtonProps, CmButtonSize, CmButtonType, CmButtonVariant } from './components/button/button.types';